run code build via a temporary docker container

+++ tests/build_code.sh
@@ -4,10 +4,9 @@
 ## Description :
 ## --
 ## Created : <2018-07-02>
-## Updated: Time-stamp: <2018-07-02 16:46:31>
+## Updated: Time-stamp: <2018-07-02 16:54:28>
 ##-------------------------------------------------------------------
 function build_code {
-    cd ..
     go get -d github.com/oratos/out_syslog/...
     go build -buildmode c-shared -o tests/out_syslog.so github.com/oratos/out_syslog/cmd
 }
@@ -16,19 +15,19 @@ function run_container {
     local container_name="go-build"
     cd ..
     # Base Image: https://hub.docker.com/r/library/golang/
-    if docker ps | grep "$container_name" >/dev/null 2>&1; then
+    if docker ps -a | grep "$container_name" >/dev/null 2>&1; then
         echo "Delete existing container: $container_name"
         docker stop "$container_name" || docker stop "$container_name" || true
         docker rm "$container_name"
     fi
 
     echo "Run container($container_name) to build the code"
-    docker run --rm -t -d -h "$container_name" --name "$container_name" \
+    docker run -t -d -h "$container_name" --name "$container_name" \
            -v "${PWD}/cmd:/go/cmd" -v "${PWD}/pkg:/go/pkg" \
            -v "${PWD}/tests:/go/tests" \
-           golang:1.10.3 bash -c "/go/tests/build_code.sh build_code"
+           golang:1.10.3 bash -c "cd /go && tests/build_code.sh build_code"
 
-    echo "To check status, run: docker logs $container_name"
+    echo "To check status, run: docker logs -f $container_name"
 }
 
 action=${1:-run_container}
